Clean up Levenberg Marquardt optimizer.
