+++ config/trajectory_editing_levenberg_marquardt.yml
@@ -7,6 +7,6 @@ dno:
   num_opt_steps: 10
   optimizer: LevenbergMarquardt # Adam, SGD, LBFGS, LevenbergMarquardt, ...
   lr: 0.5
-  enable_profiler: True
+  enable_profiler: False
   levenbergMarquardt:
     attempts_per_step: 10
